+++ src/main/java/com/phantomwing/rusticdelight/RusticDelight.java
@@ -85,7 +85,8 @@ public class RusticDelight {
                 ModItems.COTTON_BOLL.get(),
                 ModItems.BELL_PEPPER_SEEDS.get(),
                 ModItems.COTTON_SEEDS.get(),
-                ModItems.COFFEE_BEANS.get());
+                ModItems.COFFEE_BEANS.get()
+        );
 
         newWantedItems.addAll(Villager.WANTED_ITEMS);
         Villager.WANTED_ITEMS = ImmutableSet.copyOf(newWantedItems);