+++ src/main/java/com/phantomwing/rusticdelight/datagen/ModItemModelProvider.java
@@ -74,6 +74,7 @@ public class ModItemModelProvider extends ItemModelProvider {
         simpleBlock(ModBlocks.COTTON_SEEDS_BAG);
         simpleBlock(ModBlocks.BELL_PEPPER_SEEDS_BAG);
         simpleBlock(ModBlocks.COFFEE_BEANS_BAG);
+        simpleBlock(ModBlocks.ROASTED_COFFEE_BEANS_BAG);
         simpleBlock(ModBlocks.COTTON_BOLL_CRATE);
         simpleBlock(ModBlocks.BELL_PEPPER_GREEN_CRATE);
         simpleBlock(ModBlocks.BELL_PEPPER_YELLOW_CRATE);