+++ src/main/java/com/phantomwing/rusticdelight/tags/CommonTags.java
@@ -22,10 +22,11 @@ public class CommonTags {
     public static final TagKey<Item> FOODS_ONION = commonItemTag("foods/onion");
     public static final TagKey<Item> FOODS_TOMATO = commonItemTag("foods/tomato");
     public static final TagKey<Item> FOODS_BELL_PEPPER = commonItemTag("foods/bell_pepper");
+    public static final TagKey<Item> FOODS_COOKIE = commonItemTag("foods/cookie");
 
     public static final TagKey<Item> CROPS_COTTON = commonItemTag("crops/cotton");
     public static final TagKey<Item> CROPS_BELL_PEPPER = commonItemTag("crops/bell_pepper");
-    public static final TagKey<Item> CROPS_COFFEE = commonItemTag("crops/coffee");
+    public static final TagKey<Item> CROPS_COFFEE_BEANS = commonItemTag("crops/coffee_beans");
     public static final TagKey<Item> CROPS_RICE = commonItemTag("crops/rice");
 
     public static final TagKey<Item> TOOLS_KNIFE = commonItemTag("tools/knife");