+++ changelog.md
@@ -1,6 +1,19 @@
 # 1.3.1
 ### Additions
+- Added Bag of Roasted Coffee Beans
 - Added German translations
+- Added Japanese translations (thank you EndilCrafter!)
+- Added Taiwanese Mandarin translations (thank you Lobster0228!)
+- Added Cherry Blossom Cookie to the `c/foods/cookie` tag
+
+### Changes
+- Golden Coffee Beans can now also be crafted with Roasted Coffee Beans
+- Roasted Coffee Beans can now be crafted into Brown Dye
+- Made the following items compostable
+  - Cherry Blossom Cookie
+  - Cherry Blossom Cheesecake
+  - Slice of Cherry Blossom Cheesecake
+- Updated textures for Coffee Beans, Roasted Coffee Beans, Bag of Coffee Beans
 
 
 # 1.3.0